modules/ai_intelligence/pqn_alignment/requirements.txt
numpy>=1.23
pyyaml>=6.0
